Only add popup if rect contains markers
We may need to check if we need a popup to exist upon modifying the markers on the map with subsequent data loads; should we instead make an empty popup? Or, does it later detect that no popup exists and creates one, or does it flush all popups and regenerate them on new data being loaded?